Removed tuning library to separate repository
Updated benchmarks to link against a shared tuning library

This involves changes to several files including the makefile.
To generate a makefile, please use gen-makefile.sh and enter
the path to the tuning library repository.
Once this is done, the script should generate a makefile that is
specifically configured to the way you have things set up.

.gitignore has been updated to ignore 'Makefile'